+++ flang/lib/Lower/CallInterface.cpp
@@ -929,13 +929,15 @@ private:
       PassEntityBy passBy = PassEntityBy::BaseAddress;
       Property prop = Property::BaseAddress;
       if (isValueAttr) {
+        bool isBuiltinCptrType = fir::isa_builtin_cptr_type(type);
         if (isBindC || (!type.isa<fir::SequenceType>() &&
                         !obj.attrs.test(Attrs::Optional) &&
-                        dynamicType.category() !=
-                            Fortran::common::TypeCategory::Derived)) {
+                        (dynamicType.category() !=
+                             Fortran::common::TypeCategory::Derived ||
+                         isBuiltinCptrType))) {
           passBy = PassEntityBy::Value;
           prop = Property::Value;
-          if (fir::isa_builtin_cptr_type(type)) {
+          if (isBuiltinCptrType) {
             auto recTy = type.dyn_cast<fir::RecordType>();
             mlir::Type fieldTy = recTy.getTypeList()[0].second;
             passType = fir::ReferenceType::get(fieldTy);